move : 서비스 테스트 패키지 분류 추가
